feat: 위시 관련 인증 로직 추가

+++ backend/src/main/java/com/pickeat/backend/wish/application/WishService.java
@@ -3,6 +3,7 @@ package com.pickeat.backend.wish.application;
 import com.pickeat.backend.global.exception.BusinessException;
 import com.pickeat.backend.global.exception.ErrorCode;
 import com.pickeat.backend.restaurant.domain.FoodCategory;
+import com.pickeat.backend.room.domain.repository.RoomUserRepository;
 import com.pickeat.backend.wish.application.dto.request.WishRequest;
 import com.pickeat.backend.wish.application.dto.response.WishResponse;
 import com.pickeat.backend.wish.domain.Wish;
@@ -20,14 +21,16 @@ public class WishService {
 
     private final WishListRepository wishListRepository;
     private final WishRepository wishRepository;
+    private final RoomUserRepository roomUserRepository;
 
     @Transactional
     public WishResponse createWish(
             Long wishListId,
-            WishRequest request
+            WishRequest request,
+            Long userId
     ) {
-        //TODO: 방 도메인이 완성되면 현재 요청자가 현재 위시 리스트가 포함된 방의 참가지인지 검증 필요(2025-08-1, 금, 16:15)
         WishList wishList = getWishList(wishListId);
+        validateUserAccessToRoom(wishList.getRoomId(), userId);
         Wish wish = new Wish(
                 request.name(),
                 FoodCategory.getCategoryNameBy(request.category()),
@@ -40,9 +43,10 @@ public class WishService {
     }
 
     @Transactional
-    public void deleteWish(Long wishId) {
+    public void deleteWish(Long wishId, Long userId) {
         Wish wish = getWish(wishId);
-        //TODO: 방 도메인이 완성되면 현재 요청자가 현재 위시 리스트가 포함된 방의 참가지인지 검증 필요(2025-08-1, 금, 16:15)
+        WishList wishList = wish.getWishList();
+        validateUserAccessToRoom(wishList.getRoomId(), userId);
         //TODO: 위시 삭제시 위시 이미지 제거  (2025-08-4, 월, 17:59)
         wishRepository.delete(wish);
     }
@@ -55,5 +59,11 @@ public class WishService {
     private Wish getWish(Long wishId) {
         return wishRepository.findById(wishId)
                 .orElseThrow(() -> new BusinessException(ErrorCode.WISH_NOT_FOUND));
+    }
+
+    private void validateUserAccessToRoom(Long roomId, Long userId) {
+        if (!roomUserRepository.existsByRoomIdAndUserId(roomId, userId)) {
+            throw new BusinessException(ErrorCode.WISH_ACCESS_DENIED);
+        }
     }
 }
